Poprawa rysowania funkcji
naprawiłem rysowanie funkcji ( rysują sie tak jak na wykresie i skalują sie też dobrze ) poprawiłem również oznaczenia na osi y ponieważ były odwrotnie

+++ FuncDraw/FuncDraw/PointsGenerator.cs
@@ -11,7 +11,7 @@ namespace FuncDraw
         
         public List<Point> points = new List<Point>();
         
-        public void AddPoint(int x, int y)
+        public void AddPoint(double x, double y)
         {
             points.Add(new Point(x, y));
         }
